Partially operational main game (physics)
Completed:
- Basic motion
- Jumping and gravity
- Improved jumping restrictions
- Confirmed velocity modifier operation
- Restricted to CVPM

In development:
- Appropriate functionality of ice and honey
- Preventing players from moving horizontally through objects

+++ src/VelocityModifier.java
@@ -21,7 +21,7 @@ public class VelocityModifier extends Item{
 		return xyDir;
 	}
 	
-	public int[] getSpeed() {
+	public double[] getSpeed() {
 		return null;
 	}
